// src/services/proxy/generated/models/Book.ts
export interface Book {
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    isbn?: string;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    synopsis?: string;
    /**
     * 
     * @type {Date}
     * @memberof Book
     */
    publicationDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    averageRating?: number;
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    stock?: number;
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    readingAgeGropuId?: number;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    readingAgeGroupName?: string;
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    languageId?: number;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    languageName?: string;
    /**
     * 
     * @type {number}
     * @memberof Book
     */
    formatId?: number;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    formatName?: string;
    /**
     * 
     * @type {Array<Author>}
     * @memberof Book
     */
    authors?: Array<Author>;
    /**
     * 
     * @type {Array<Subject>}
     * @memberof Book
     */
    subjects?: Array<Subject>;
    /**
     * 
     * @type {Array<LiteraryGenre>}
     * @memberof Book
     */
    literaryGenres?: Array<LiteraryGenre>;
}