// src/services/proxy/generated/models/Author.ts
/* tslint:disable */
/* eslint-disable */
/**
 * The Golden Book API
 * API for a book-selling business
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thegoldenbook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Author
 */
export interface Author {
    /**
     * 
     * @type {number}
     * @memberof Author
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Author
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Author
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof Author
     */
    secondLastName?: string;
    /**
     * 
     * @type {Date}
     * @memberof Author
     */
    dateOfBirth?: Date;
}

/**
 * Check if a given object implements the Author interface.
 */
export function instanceOfAuthor(value: object): value is Author {
    return true;
}

export function AuthorFromJSON(json: any): Author {
    return AuthorFromJSONTyped(json, false);
}

export function AuthorFromJSONTyped(json: any, ignoreDiscriminator: boolean): Author {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'secondLastName': json['secondLastName'] == null ? undefined : json['secondLastName'],
        'dateOfBirth': json['dateOfBirth'] == null ? undefined : (new Date(json['dateOfBirth'])),
    };
}

export function AuthorToJSON(json: any): Author {
    return AuthorToJSONTyped(json, false);
}

export function AuthorToJSONTyped(value?: Author | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'lastName': value['lastName'],
        'secondLastName': value['secondLastName'],
        'dateOfBirth': value['dateOfBirth'] == null ? undefined : ((value['dateOfBirth']).toISOString()),
    };
}