// src/services/proxy/generated/models/Subject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * The Golden Book API
 * API for a book-selling business
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thegoldenbook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Subject
 */
export interface Subject {
    /**
     * 
     * @type {number}
     * @memberof Subject
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Subject
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof Subject
     */
    parentId?: number;
    /**
     * 
     * @type {string}
     * @memberof Subject
     */
    parentName?: string;
}

/**
 * Check if a given object implements the Subject interface.
 */
export function instanceOfSubject(value: object): value is Subject {
    return true;
}

export function SubjectFromJSON(json: any): Subject {
    return SubjectFromJSONTyped(json, false);
}

export function SubjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): Subject {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'parentId': json['parentId'] == null ? undefined : json['parentId'],
        'parentName': json['parentName'] == null ? undefined : json['parentName'],
    };
}

export function SubjectToJSON(json: any): Subject {
    return SubjectToJSONTyped(json, false);
}

export function SubjectToJSONTyped(value?: Subject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'parentId': value['parentId'],
        'parentName': value['parentName'],
    };
}