// src/services/proxy/generated/models/Review.ts
export interface Review {
    /**
     * 
     * @type {number}
     * @memberof Review
     */
    bookId?: number;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    bookName?: string;
    /**
     * 
     * @type {number}
     * @memberof Review
     */
    authorId?: number;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    authorName?: string;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    authorLastName?: string;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    authorSecondLastName?: string;
    /**
     * 
     * @type {number}
     * @memberof Review
     */
    userId?: number;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    nickname?: string;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    userName?: string;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    userLastName?: string;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    userSecondLastName?: string;
    /**
     * 
     * @type {number}
     * @memberof Review
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    subject?: string;
    /**
     * 
     * @type {string}
     * @memberof Review
     */
    body?: string;
    /**
     * 
     * @type {Date}
     * @memberof Review
     */
    publicationDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof Review
     */
    languageId?: number;
}